Create a test for createKing function and organize the project
