+++ .eslintrc.json
@@ -4,10 +4,10 @@
     "es2021": true,
     "node": true
   },
-  "extends": ["xo", "prettier"],
+  "extends": ["xo", "prettier", "jest"],
   "overrides": [
     {
-      "extends": ["xo-typescript", "prettier"],
+      "extends": ["xo-typescript", "prettier", "jest"],
       "files": ["*.ts", "*.tsx"],
       "rules": {
         "@typescript-eslint/consistent-type-definitions": ["error", "interface"]
