+++ package.json
@@ -7,6 +7,7 @@
     "build": "tsc",
     "build:dev": "tsc -w",
     "start": "node .",
+    "test": "jest --verbose",
     "prepare": "husky"
   },
   "keywords": [],
@@ -16,6 +17,7 @@
   "devDependencies": {
     "@commitlint/cli": "^19.3.0",
     "@commitlint/config-conventional": "^19.2.2",
+    "@types/jest": "^29.5.12",
     "@typescript-eslint/eslint-plugin": "^7.18.0",
     "@typescript-eslint/parser": "^7.18.0",
     "eslint": "^8.57.0",
@@ -23,6 +25,8 @@
     "eslint-config-xo": "^0.45.0",
     "eslint-config-xo-typescript": "^5.0.0",
     "husky": "^9.1.4",
+    "jest": "^29.7.0",
+    "ts-jest": "^29.2.3",
     "typescript": "^5.5.4"
   }
 }
