+++ src/characters-factory-functions.ts
@@ -5,30 +5,8 @@ import type {
   DexterityRange,
   FidelityRange,
   Fighter,
-  King,
   Squire,
 } from "./types";
-
-export const createKing = (
-  { name, surname, age }: mandatoryData,
-  yearsOfReign: number
-): King => {
-  const king: King = {
-    name,
-    surname,
-    age,
-    isAlive: true,
-    yearsOfReign,
-    talk() {
-      return "Vais a morir todos";
-    },
-    die() {
-      king.isAlive = false;
-    },
-  };
-
-  return king;
-};
 
 export const createFighter = (
   { name, surname, age }: mandatoryData,
