+++ src/characters/king/type.ts
@@ -1,0 +1,6 @@
+import { Character } from "../../types";
+
+export interface King extends Character {
+  yearsOfReign: number;
+  talk: () => "Vais a morir todos";
+}
